refactor(settings): extract VoiceSettingsScreen to its own file and clean up SettingsHomeScreen
