+++ lib/features/settings/presentation/voice_settings_screen.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+import '../widgets/voice_settings_section.dart';
+
+class VoiceSettingsScreen extends StatelessWidget {
+  const VoiceSettingsScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Voice Settings')),
+      body: const VoiceSettingsSection(),
+    );
+  }
+}
